requirements.txt
anthropic>=0.40.0
fastapi>=0.115.0
uvicorn>=0.32.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0
